Bump Version Format Fix Warnings

+++ src/particle.rs
@@ -1,5 +1,5 @@
 use glam::Vec2;
-use rand::{Rng, SeedableRng};
+use rand::Rng;
 
 use crate::physics::PhysicsModule;
 
@@ -17,9 +17,9 @@ pub fn generate_particles(queue: &wgpu::Queue, physics_module: &PhysicsModule, n
     let mut rng = rand::thread_rng();
 
     // Generate Chunks of Random Particles
-    for c in 0..num_particles as u64 / 128 {
+    for c in 0..num_particles / 128 {
         let chunk = Vec2::new(rng.gen_range(-20f32..=20f32), rng.gen_range(-20f32..=20f32));
-        for p in 0..128 as u64 {
+        for p in 0..128 {
             let dir = Vec2::new(rng.gen_range(-1f32..=1f32), rng.gen_range(-1f32..=1f32));
             let d = rng.gen_range(0.0..=4.0);
             let particle = Particle {
@@ -29,7 +29,7 @@ pub fn generate_particles(queue: &wgpu::Queue, physics_module: &PhysicsModule, n
                 mass: 0.1,   //rng.gen_range(0.01..=0.2f32),
             };
 
-            let i = c + p * (num_particles as u64 / 128);
+            let i = c + p * (num_particles / 128);
             queue.write_buffer(
                 physics_module.current_buffer(),
                 i * 24,
